Remove the async-support crate
