+++ src/sensors/mod.rs
@@ -1,9 +1,9 @@
+use crate::executor;
 use crate::futures;
 use crate::result::TockResult;
 use crate::syscalls;
 use core::cell::Cell;
 use core::convert::From;
-use core::executor;
 use core::fmt;
 use core::mem;
